apps/api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
pydantic==2.7.2
pydantic-settings==2.2.1
email-validator==2.2.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
aiofiles==23.2.1
boto3==1.34.107
faster-whisper==1.1.0
pytest==8.2.1
httpx==0.27.0
requests==2.32.3
Cython
packaging
